add custom routes, options
Adding custom routes and add backup:run $options.

+++ src/routes/backpack/backupmanager.php
@@ -15,8 +15,10 @@ Route::group([
     'prefix'     => config('backpack.base.route_prefix', 'admin'),
     'middleware' => ['web', 'admin'],
 ], function () {
-    Route::get('backup', 'BackupController@index');
-    Route::put('backup/create', 'BackupController@create');
-    Route::get('backup/download/{file_name?}', 'BackupController@download');
-    Route::delete('backup/delete/{file_name?}', 'BackupController@delete')->where('file_name', '(.*)');
+    if (config('backpack.backup.setup_backup_routes')) {
+        Route::get('backup', 'BackupController@index');
+        Route::put('backup/create', 'BackupController@create');
+        Route::get('backup/download/{file_name?}', 'BackupController@download');
+        Route::delete('backup/delete/{file_name?}', 'BackupController@delete')->where('file_name', '(.*)');
+    }
 });
